More work on Score Framework
All images will be reworked into a compatible format by the next commit.
The Score framework has seen some changes and the ability to change the
image. Bugs still exist with the image display. There is a white
backround behind the images, which should be fixed by the next commit.