Loads of stuff added.
Collision detection was improved, not full proof though. Textures for
the ball and paddles have been added as well as sounds. The restart
button has been tweaked and the game is in a releasable state. This was
my first ever complete game. It took me 3 days to do, with on and off
sessions. Tetris is next!